app/src/main/java/com/dedesaepulloh/submissionbajp/ui/movies/MoviesViewModel.kt
package com.dedesaepulloh.submissionbajp.ui.movies

import androidx.lifecycle.LiveData
import androidx.lifecycle.ViewModel
import androidx.paging.PagedList
import com.dedesaepulloh.submissionbajp.data.source.CatalogRepository
import com.dedesaepulloh.submissionbajp.data.source.local.entity.MovieEntity
import com.dedesaepulloh.submissionbajp.vo.Resource

class MoviesViewModel(private val catalogRepository: CatalogRepository) : ViewModel() {
    fun getMovies(): LiveData<Resource<PagedList<MovieEntity>>> =
        catalogRepository.getMoviePopular()
}